Simplify and rearchitect the whole thing
- States and actions are not abstract classes anymore.
They just take names which are strings. This elminates the
need to keep extending them since that doesn't provide
any benefit to the consumer.

- States contain substates within them instead of having the
concept of parent and child machines. This is simpler to
grok as there is a single state machine with possible substates.
i.e. it's a hierachical finite state machine.

-  Every state now gets an optional entry and exit
handler.

-  Instead of transitions, we have have trigger handlers
now. This means consumer is reponsible to transitioning state
on event or doing whatever else they want.