Take bintray key from environment variable instead of keystore.gradle
We can't use apply from 'keystore.gradle' and not commit keystore.gradle
to git since CI won't be able to find it if it's not in git. This means
that the only option to not commit credentials to git and still have it
available is to make it an environment variable.