Fix duplicate trigger comparison (#2)
With the current implementation, exceptions aren't always thrown if a duplicate trigger handler is added. This is because the states are being compared by reference instead of by the contents of the object. By overriding the .equals(Any?) method of State this problem is now fixed. Note that onEnter and onExit aren't used in the comparison as these shouldn't be considered (hence the use of manually overriding .equals(Any?) and hashCode() compared to using a data class).